--- a/bessctl/conf/perftest/phy_forward.bess
+++ b/bessctl/conf/perftest/phy_forward.bess
@@ -50,5 +50,5 @@ for i in range(num_ports):
         QueueInc(name=qinc_name, port=ports[inc_port_id], qid=j) \
         -> QueueOut(name=qout_name, port=ports[out_port_id], qid=j)
 
-        bess.attach_task(qinc_name, 0, wid=next_wid)
+        bess.attach_module(qinc_name, wid=next_wid)
         next_wid = (next_wid + 1) % num_cores