--- a/bessctl/conf/samples/multicore.bess
+++ b/bessctl/conf/samples/multicore.bess
@@ -28,8 +28,8 @@ src1::Source() -> sink
 #attach src0:0 => worker:0 
 #attach src1:0 => worker:1
 
-bess.attach_task("src0", 0, wid=0)
-bess.attach_task("src1", 0, wid=1)
+bess.attach_module('src0', wid=0)
+bess.attach_module('src1', wid=1)
 
 # Future work
 #