--- a/bessctl/conf/samples/tc/ratelimit.bess
+++ b/bessctl/conf/samples/tc/ratelimit.bess
@@ -1,47 +1,35 @@
 # Check out "show tc" and "monitor tc" commands
 
-bess.add_tc('rr', policy='round_robin', priority=0)
-
 # for every second, src0 is scheduled 100000 times 
 src0::Source() -> Sink()
 bess.add_tc('schedule_limit',
-            parent='rr',
             policy='rate_limit',
             resource='count',
             limit={'count': 100000})
-bess.add_tc('schedule_leaf', policy='leaf', parent='schedule_limit')
-bess.attach_task(src0.name, tc='schedule_leaf')
+bess.attach_module(src0.name, 'schedule_limit')
 
 # up to 100Mhz
 src1::Source() -> Sink()
 bess.add_tc('cycle_limit',
-            parent='rr',
             policy='rate_limit',
             resource='cycle',
             limit={'cycle': 100000000})
-bess.add_tc('cycle_leaf', policy='leaf', parent='cycle_limit')
-bess.attach_task(src1.name, tc='cycle_leaf')
+bess.attach_module(src1.name, 'cycle_limit')
 
 # 10M packets / sec
 src2::Source() -> Sink()
 bess.add_tc('packet_limit',
-            parent='rr',
             policy='rate_limit',
             resource='packet',
             limit={'packet': 10000000})
-bess.add_tc('packet_leaf', policy='leaf', parent='packet_limit')
-bess.attach_task(src2.name, tc='packet_leaf')
+bess.attach_module(src2.name, 'packet_limit')
 
 # Gigabits per second
 src3::Source() -> Sink()
 bess.add_tc('bit_limit',
-            parent='rr',
             policy='rate_limit',
             resource='bit',
             limit={'bit': 1000000000})
-bess.add_tc('bit_leaf', policy='leaf', parent='bit_limit')
-bess.attach_task(src3.name, tc='bit_leaf')
+bess.attach_module(src3.name, 'bit_limit')
 
 src4::Source() -> Sink()
-bess.add_tc('unlimited', parent='rr', policy='leaf')
-bess.attach_task(src4.name, tc='unlimited')