--- a/bessctl/conf/testing/run_module_tests.bess
+++ b/bessctl/conf/testing/run_module_tests.bess
@@ -77,7 +77,7 @@ def load_test(filename):
             CUSTOM_TEST_FUNCTIONS]
 
 def monitor_task(module, wid):
-  bess.attach_task(module.name, 0, wid=wid)
+  bess.attach_module(module.name, wid=wid)
 
 ## RUN TEST LOOP ##
 path = os.path.dirname(os.path.realpath(sys.argv[0]))