--- a/core/task.h
+++ b/core/task.h
@@ -18,8 +18,11 @@ struct task_result {
   uint64_t bits;
 };
 
+// Stores the arguments of a task created by a module.
 class ModuleTask {
  public:
+  // Doesn't take ownership of 'arg' and 'c'.  'c' can be null and it
+  // can be changed later with SetTC()
   ModuleTask(void *arg, bess::LeafTrafficClass<Task> *c)
       : arg_(arg), c_(c) {};
 
@@ -38,20 +41,25 @@ class ModuleTask {
   }
 
  private:
-  void *arg_;
-  bess::LeafTrafficClass<Task> *c_;
+  void *arg_; // Auxiliary value passed to Module::RunTask().
+  bess::LeafTrafficClass<Task> *c_; // Leaf TC associated with this task.
 };
 
+// Functor used by a leaf in a Worker's Scheduler to run a task in a module.
 class Task {
  public:
+  // When this task is scheduled it will execute 'm' with 'arg'.
+  // When the associated leaf is created/destroyed, 't' will be updated.
   Task(Module *m, void *arg, ModuleTask *t) : module_(m), arg_(arg), t_(t) {};
 
+  // Called when the leaf that owns this task is destroyed.
   void Detach() {
     if (t_) {
       t_->SetTC(nullptr);
     }
   }
 
+  // Called when the leaf that owns this task is created.
   void Attach(bess::LeafTrafficClass<Task> *c) {
     if (t_) {
       t_->SetTC(c);
@@ -61,8 +69,11 @@ class Task {
   struct task_result operator()(void);
 
  private:
+  // Used by operator().
   Module *module_;
   void *arg_;
+
+  // Used to notify a module that a leaf is being created/destroyed.
   ModuleTask *t_;
 };