With this pull request LeafTrafficClass can only execute a single Task (more details in the last commit message).

I have also implemented (in a separate commit) the tree view for show tc.  Sample output:

```
<worker 0>
  +-- main                     priority
      +-- rl                   rate_limit          priority: 5         limits: 40.000 Mbps
      |   +-- !leaf_src0:0     leaf
      +-- rest                 round_robin         priority: 100
          +-- 1                weighted_fair
          |   +-- 11           round_robin         share: 5
          |   +-- 12           round_robin         share: 6
          +-- 2                round_robin
```

Suggestions on how to improve are welcome!