@ddiproietto 2017-04-07
I didn't realize that `return_with_error()` returns Status::OK, so there's no way for the callers of `FindTC()` to understand if the TC has been found.  Luckily g++-4.8 reported a warning.

I'm pushing a new commit