@ddiproietto 2017-04-11
These are the results of running traffic_class_bench, before (f249f1218e48) and after(d580cdc4d21a) the refactor:

After the refactor we lose the possibility to run multiple tasks on a single leaf, so we lose the TCLeaf/TCRunTaskOnce benchmark.  Its results should be compared with TCRoundRobin/TCScheduleOnce.

Before:

```
----------------------------------------------------------------------------------
Benchmark                                           Time           CPU Iterations
----------------------------------------------------------------------------------
TCWeightedFair/TCScheduleOnceCount/4/0             26 ns         26 ns   27161882    37.016M items/s
TCWeightedFair/TCScheduleOnceCount/8/0             27 ns         27 ns   25522986   34.8511M items/s
TCWeightedFair/TCScheduleOnceCount/16/0            29 ns         29 ns   23911187   32.5962M items/s
TCWeightedFair/TCScheduleOnceCount/32/0            31 ns         31 ns   22438633   30.4667M items/s
TCWeightedFair/TCScheduleOnceCount/64/0            34 ns         34 ns   20369406    27.678M items/s
TCWeightedFair/TCScheduleOnceCount/128/0           42 ns         42 ns   16509520   22.4628M items/s
TCWeightedFair/TCScheduleOnceCount/256/0           51 ns         51 ns   13015095   18.7587M items/s
TCWeightedFair/TCScheduleOnceCount/512/0           59 ns         59 ns   11285487   16.2775M items/s
TCWeightedFair/TCScheduleOnceCount/1024/0          68 ns         68 ns    9685476   13.9764M items/s
TCWeightedFair/TCScheduleOnceCount/2048/0          76 ns         76 ns    8674331   12.5198M items/s
TCWeightedFair/TCScheduleOnceCount/4096/0          81 ns         81 ns    8445245   11.8019M items/s
TCWeightedFair/TCScheduleOnceCount/8192/0          87 ns         87 ns    8301092   10.9881M items/s
TCWeightedFair/TCScheduleOnceCount/16384/0        124 ns        124 ns    6167515   7.69309M items/s
TCWeightedFair/TCScheduleOnceCount/32768/0        145 ns        145 ns    5440937    6.5985M items/s
TCWeightedFair/TCScheduleOnceCount/65536/0        142 ns        142 ns    5243198    6.6991M items/s
TCWeightedFair_BigO                              7.79 lgN       7.79 lgN
TCWeightedFair_RMS                                 19 %         19 %
TCWeightedFair/TCScheduleOnceCycle/4/1             29 ns         29 ns   24127392   32.9158M items/s
TCWeightedFair/TCScheduleOnceCycle/8/1             32 ns         32 ns   21705680   29.6339M items/s
TCWeightedFair/TCScheduleOnceCycle/16/1            39 ns         39 ns   18139407   24.7293M items/s
TCWeightedFair/TCScheduleOnceCycle/32/1            48 ns         48 ns   13685969   19.9763M items/s
TCWeightedFair/TCScheduleOnceCycle/64/1            54 ns         54 ns   12488681   17.7202M items/s
TCWeightedFair/TCScheduleOnceCycle/128/1           61 ns         61 ns   10881437   15.5772M items/s
TCWeightedFair/TCScheduleOnceCycle/256/1           68 ns         68 ns    9723510   13.9882M items/s
TCWeightedFair/TCScheduleOnceCycle/512/1           78 ns         78 ns    8674833   12.2925M items/s
TCWeightedFair/TCScheduleOnceCycle/1024/1          89 ns         89 ns    7579622   10.7653M items/s
TCWeightedFair/TCScheduleOnceCycle/2048/1         102 ns        102 ns    6926232   9.34808M items/s
TCWeightedFair/TCScheduleOnceCycle/4096/1         109 ns        109 ns    6233068   8.74974M items/s
TCWeightedFair/TCScheduleOnceCycle/8192/1         125 ns        125 ns    5559447   7.63817M items/s
TCWeightedFair/TCScheduleOnceCycle/16384/1        187 ns        187 ns    3770932   5.10243M items/s
TCWeightedFair/TCScheduleOnceCycle/32768/1        231 ns        231 ns    3091050   4.13422M items/s
TCWeightedFair/TCScheduleOnceCycle/65536/1        289 ns        289 ns    2471269   3.29894M items/s
TCWeightedFair_BigO                             12.16 lgN      12.16 lgN
TCWeightedFair_RMS                                 35 %         35 %
TCRoundRobin/TCScheduleOnce/1                      19 ns         19 ns   37078034   50.6769M items/s
TCRoundRobin/TCScheduleOnce/4                      19 ns         19 ns   36847804   50.7513M items/s
TCRoundRobin/TCScheduleOnce/8                      19 ns         19 ns   36370365    49.717M items/s
TCRoundRobin/TCScheduleOnce/16                     19 ns         19 ns   36757877   50.2725M items/s
TCRoundRobin/TCScheduleOnce/32                     19 ns         19 ns   36771580   50.4955M items/s
TCRoundRobin/TCScheduleOnce/64                     19 ns         19 ns   36937390   50.5279M items/s
TCRoundRobin/TCScheduleOnce/128                    20 ns         20 ns   36117373   48.7172M items/s
TCRoundRobin/TCScheduleOnce/256                    19 ns         19 ns   35935740   49.4276M items/s
TCRoundRobin/TCScheduleOnce/512                    20 ns         20 ns   35415743   48.3292M items/s
TCRoundRobin/TCScheduleOnce/1024                   20 ns         20 ns   35326163   48.3509M items/s
TCRoundRobin/TCScheduleOnce/2048                   20 ns         20 ns   35260106   47.9361M items/s
TCRoundRobin/TCScheduleOnce/4096                   20 ns         20 ns   35846544   48.0754M items/s
TCRoundRobin/TCScheduleOnce/8192                   20 ns         20 ns   35292123   48.3978M items/s
TCRoundRobin/TCScheduleOnce/16384                  25 ns         25 ns   28244831    38.316M items/s
TCRoundRobin/TCScheduleOnce/32768                  28 ns         28 ns   25485603   34.6618M items/s
TCRoundRobin/TCScheduleOnce/65536                  28 ns         28 ns   24752426   33.9328M items/s
TCRoundRobin_BigO                               20.74 (1)      20.74 (1)
TCRoundRobin_RMS                                   15 %         15 %
TCLeaf/TCRunTaskOnce/1                             19 ns         19 ns   37081273   50.7385M items/s
TCLeaf/TCRunTaskOnce/4                             26 ns         26 ns   27302985   148.776M items/s
TCLeaf/TCRunTaskOnce/8                             44 ns         44 ns   16102843   175.347M items/s
TCLeaf/TCRunTaskOnce/16                            66 ns         66 ns   10301169   230.111M items/s
TCLeaf/TCRunTaskOnce/32                           120 ns        120 ns    5796069   255.216M items/s
TCLeaf/TCRunTaskOnce/64                           212 ns        212 ns    3296937   288.551M items/s
TCLeaf/TCRunTaskOnce/128                          395 ns        395 ns    1770811   309.179M items/s
TCLeaf/TCRunTaskOnce/256                          763 ns        763 ns     890952   320.145M items/s
TCLeaf/TCRunTaskOnce/512                         1597 ns       1597 ns     435230   305.794M items/s
TCLeaf/TCRunTaskOnce/1024                        3273 ns       3273 ns     213402   298.379M items/s
TCLeaf/TCRunTaskOnce/2048                        6508 ns       6508 ns     102269   300.123M items/s
TCLeaf/TCRunTaskOnce/4096                       12966 ns      12966 ns      52896   301.265M items/s
TCLeaf/TCRunTaskOnce/8192                       25746 ns      25745 ns      27192   303.453M items/s
TCLeaf/TCRunTaskOnce/16384                      51372 ns      51372 ns      13651   304.152M items/s
TCLeaf/TCRunTaskOnce/32768                     102636 ns     102633 ns       6848   304.484M items/s
TCLeaf/TCRunTaskOnce/65536                     229958 ns     229957 ns       3064    271.79M items/s
TCLeaf_BigO                                      0.22 NlgN       0.22 NlgN
TCLeaf_RMS                                          6 %          6 %

```

After:

```
----------------------------------------------------------------------------------
Benchmark                                           Time           CPU Iterations
----------------------------------------------------------------------------------
TCWeightedFair/TCScheduleOnceCount/4/0             23 ns         23 ns   29892111   40.6927M items/s
TCWeightedFair/TCScheduleOnceCount/8/0             25 ns         25 ns   27968308   38.2001M items/s
TCWeightedFair/TCScheduleOnceCount/16/0            27 ns         27 ns   26062292   35.5311M items/s
TCWeightedFair/TCScheduleOnceCount/32/0            29 ns         29 ns   23986816   33.0294M items/s
TCWeightedFair/TCScheduleOnceCount/64/0            31 ns         31 ns   22425569   30.6062M items/s
TCWeightedFair/TCScheduleOnceCount/128/0           35 ns         34 ns   20097630   27.6455M items/s
TCWeightedFair/TCScheduleOnceCount/256/0           47 ns         47 ns   14869223   20.2776M items/s
TCWeightedFair/TCScheduleOnceCount/512/0           54 ns         54 ns   12418570   17.7798M items/s
TCWeightedFair/TCScheduleOnceCount/1024/0          58 ns         58 ns   11654107   16.3586M items/s
TCWeightedFair/TCScheduleOnceCount/2048/0          65 ns         65 ns   10540511    14.679M items/s
TCWeightedFair/TCScheduleOnceCount/4096/0          69 ns         69 ns    9825508   13.9072M items/s
TCWeightedFair/TCScheduleOnceCount/8192/0          70 ns         70 ns    9751264   13.5903M items/s
TCWeightedFair/TCScheduleOnceCount/16384/0         78 ns         78 ns    9261178   12.2278M items/s
TCWeightedFair/TCScheduleOnceCount/32768/0        106 ns        106 ns    7513232   8.96276M items/s
TCWeightedFair/TCScheduleOnceCount/65536/0        110 ns        110 ns    6923088   8.68738M items/s
TCWeightedFair_BigO                              6.11 lgN       6.11 lgN
TCWeightedFair_RMS                                 13 %         13 %
TCWeightedFair/TCScheduleOnceCycle/4/1             25 ns         25 ns   27800708    37.909M items/s
TCWeightedFair/TCScheduleOnceCycle/8/1             30 ns         30 ns   23507936   32.1672M items/s
TCWeightedFair/TCScheduleOnceCycle/16/1            37 ns         37 ns   18705178   25.6303M items/s
TCWeightedFair/TCScheduleOnceCycle/32/1            44 ns         44 ns   15792091   21.4445M items/s
TCWeightedFair/TCScheduleOnceCycle/64/1            51 ns         51 ns   13248440     18.58M items/s
TCWeightedFair/TCScheduleOnceCycle/128/1           59 ns         59 ns   11474649   16.2522M items/s
TCWeightedFair/TCScheduleOnceCycle/256/1           65 ns         65 ns   10404612   14.5876M items/s
TCWeightedFair/TCScheduleOnceCycle/512/1           73 ns         73 ns    9375102   13.1189M items/s
TCWeightedFair/TCScheduleOnceCycle/1024/1          82 ns         82 ns    8208846   11.5788M items/s
TCWeightedFair/TCScheduleOnceCycle/2048/1          94 ns         94 ns    7231423   10.1376M items/s
TCWeightedFair/TCScheduleOnceCycle/4096/1         105 ns        105 ns    6505976   9.12169M items/s
TCWeightedFair/TCScheduleOnceCycle/8192/1         115 ns        115 ns    6056582   8.25834M items/s
TCWeightedFair/TCScheduleOnceCycle/16384/1        141 ns        141 ns    5142767   6.77133M items/s
TCWeightedFair/TCScheduleOnceCycle/32768/1        198 ns        198 ns    3586947   4.81955M items/s
TCWeightedFair/TCScheduleOnceCycle/65536/1        221 ns        221 ns    3260305     4.308M items/s
TCWeightedFair_BigO                             10.37 lgN      10.37 lgN
TCWeightedFair_RMS                                 25 %         25 %
TCRoundRobin/TCScheduleOnce/1                      16 ns         16 ns   42818187   58.6596M items/s
TCRoundRobin/TCScheduleOnce/4                      16 ns         16 ns   42870732   58.7145M items/s
TCRoundRobin/TCScheduleOnce/8                      16 ns         16 ns   42831997   58.7034M items/s
TCRoundRobin/TCScheduleOnce/16                     17 ns         17 ns   42148493   57.7155M items/s
TCRoundRobin/TCScheduleOnce/32                     16 ns         16 ns   42613975   58.1497M items/s
TCRoundRobin/TCScheduleOnce/64                     16 ns         16 ns   42481744   58.3666M items/s
TCRoundRobin/TCScheduleOnce/128                    16 ns         16 ns   42591545   58.3443M items/s
TCRoundRobin/TCScheduleOnce/256                    16 ns         16 ns   42508430   58.3475M items/s
TCRoundRobin/TCScheduleOnce/512                    16 ns         16 ns   42244717    58.456M items/s
TCRoundRobin/TCScheduleOnce/1024                   16 ns         16 ns   42047911    57.905M items/s
TCRoundRobin/TCScheduleOnce/2048                   17 ns         17 ns   41623611   57.4143M items/s
TCRoundRobin/TCScheduleOnce/4096                   17 ns         17 ns   41816284   57.5247M items/s
TCRoundRobin/TCScheduleOnce/8192                   17 ns         17 ns   41495498   57.6489M items/s
TCRoundRobin/TCScheduleOnce/16384                  17 ns         17 ns   40021240   54.6238M items/s
TCRoundRobin/TCScheduleOnce/32768                  19 ns         19 ns   36457546   49.6848M items/s
TCRoundRobin/TCScheduleOnce/65536                  21 ns         21 ns   33132649   45.6203M items/s
TCRoundRobin_BigO                               16.92 (1)      16.92 (1)
TCRoundRobin_RMS                                    7 %          7 %

```